warning removed : int double
git-svn-id: https://elmar.ips.cs.tu-bs.de/svn/cbp_ws1011@385 a60ba226-48e3-df11-83e3-000476a39db3

+++ optimizer/src/Intervals.cpp
@@ -17,10 +17,10 @@ namespace detail {
 }
 
 Interval::Interval(double val)
-    : low(std::floor(val)), up(std::ceil(val)) {}
+    : low((int)std::floor(val)), up((int)std::ceil(val)) {}
 
 Interval::Interval(double low_, double up_)
-    : low(std::floor(low_)), up(std::ceil(up_)) {}
+    : low((int)std::floor(low_)), up((int)std::ceil(up_)) {}
 
 Interval Interval::world() {
     return Interval(std::numeric_limits<int>::min(), std::numeric_limits<int>::max());
